requirements.txt
bcrypt==4.0.1
fastapi==0.89.1
passlib==1.7.4
pydantic==1.10.4
python-dotenv==0.21.0
python-jose==3.3.0
python-multipart==0.0.5
psycopg2-binary==2.9.5
sqlalchemy==1.4.46
uvicorn==0.20.0